Moved json db driver to sub package
